increase fig-width to fully display bottom legend
It's a bug of timePlot that the bottom legend doesn't swap lines when it's too long.  In this case, the PM10 in the figure is cut as PM1 which is misleading. 

I believe slightly increasing fig-width could solve the problem. BUT I DIDN'T COMPILE AND CHECK THE RESULT. PLEASE BE AWARE OF THIS BEFORE MERGE.